# bin/default/META-INF/swagger/swagger.yml
openapi: 3.0.1
paths:
...
  /helloController/echo:
    post:
      operationId: test11
      requestBody:
        content:
          text/plain:
            schema:
              type: string
        required: true
      responses:
        "200":
          description: test11 200 response
          content:
            text/plain:
              schema:
                type: string
...
  /helloController/mul:
    get:
      tags:
      - greeting
      summary: Greets a person
      description: A friendly greeting is returned
      operationId: mul
      responses:
        "400":
          description: Invalid Name Supplied
        "404":
          description: Person not found
        "200":
          content:
            text/plain: {}
  /helloController/save:
    post:
      operationId: save
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/Person'
        required: true
      responses:
        "200":
          description: save 200 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Person'
components:
  schemas:
    Person:
      required:
      - name
      type: object
      properties:
        age:
          type: integer
          format: int32
        name:
          type: string